moved programsList from component to container
