+++ src/applications/gi/containers/ProgramsList.jsx
@@ -7,7 +7,7 @@ import {
   VaPagination,
   VaTextInput,
 } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
-import { formatProgramType, generateMockPrograms } from '../../utils/helpers';
+import { formatProgramType, generateMockPrograms } from '../utils/helpers';
 // import { fetchInstitutionPrograms } from '../../actions';
 
 const ProgramsList = ({ match }) => {
